Connect edges instead of vertexes
We're connecting by the edge now instead of the vertex.

Also added in the UI changes to decide if we want to use
the binary slope of the grayscale slope.

+++ point.hpp
@@ -114,7 +114,7 @@ public:
 };
 
 
-typedef int32_t point_t;
+typedef double point_t;
 typedef point<point_t> pt;
 typedef std::pair<pt,pt> edge;
